requirements.txt
fastmcp
google-api-python-client
google-auth-oauthlib
pydantic
click
PyPDF2
python-dotenv
pydantic-settings